+++ src/libYARP_dev/src/yarp/dev/MapGrid2D.cpp
@@ -223,9 +223,9 @@ void MapGrid2D::enlargeCell(XYCell cell)
     int i = cell.x;
     int j = cell.y;
     int il = cell.x > 1 ? cell.x - 1 : 0;
-    int ir = cell.x + 1<(int)(m_width) - 1 ? cell.x + 1 : (int)(m_width)-1;
+    int ir = cell.x + 1<(m_width) - 1 ? cell.x + 1 : (m_width)-1;
     int ju = cell.y > 1 ? cell.y - 1 : 0;
-    int jd = cell.y + 1<(int)(m_height) - 1 ? cell.y + 1 : (int)(m_height)-1;
+    int jd = cell.y + 1<(m_height) - 1 ? cell.y + 1 : (m_height)-1;
 
     if (m_map_flags.pixel(il, j) == MAP_CELL_FREE) m_map_flags.pixel(il, j) = MAP_CELL_ENLARGED_OBSTACLE;
     if (m_map_flags.pixel(ir, j) == MAP_CELL_FREE) m_map_flags.pixel(ir, j) = MAP_CELL_ENLARGED_OBSTACLE;
